Created report controller methods for diet and fitness

+++ controllers/diet.js
@@ -2,6 +2,49 @@ const ErrorResponse = require('../utils/errorResponse');
 const asyncHandler = require('../middleware/async');
 const prisma = require('../utils/prismaClient');
 
+// @desc      Get Top Diets From Logs w/ Aggregated Data Based on User
+// @route     GET /api/v1/diet/report/
+// @access    Private
+exports.getDietReport = asyncHandler(async (req, res, next) => {
+  const id = req.params.user_id;
+
+  const dietLogs = await prisma.diet_logs.groupBy({
+    by: ['meal'],
+    where: {
+      user_id: id,
+    },
+    count: {
+      meal: true,
+    },
+  });
+
+  const dietIds = dietLogs.map((log) => {
+    return log.meal;
+  });
+
+  const meals = await prisma.meals.findMany({
+    where: {
+      id: {
+        in: dietIds,
+      },
+    },
+  });
+
+  const mealReport = meals.map((meal) => {
+    return {
+      id: meal.id,
+      user_id: meal.user_id,
+      created_at: meal.created_at,
+      name: meal.name,
+      description: meal.description,
+      calories: parseInt(meal.calories),
+      meal_count: dietLogs.find((log) => log.meal === meal.id).count.meal,
+    };
+  });
+
+  res.status(200).json({ success: true, data: mealReport });
+});
+
 // @desc      Get ALL Diet Logs
 // @route     GET /api/v1/diet/logs/
 // @access    Private
@@ -103,6 +146,19 @@ exports.createDiet = asyncHandler(async (req, res, next) => {
 exports.updateDietByDietId = asyncHandler(async (req, res, next) => {
   const id = req.params.diet_log_id;
 
+  // Validate User Owns Resource
+  const checkFitnessLog = await prisma.diet_logs.findUnique({
+    where: {
+      id: id,
+    },
+  });
+
+  if (checkFitnessLog.user_id !== req.user.id) {
+    return next(
+      new ErrorResponse('Not authorized to update this resource', 403)
+    );
+  }
+
   const updatedDietLog = await prisma.diet_logs.update({
     where: {
       id: id,
@@ -137,6 +193,19 @@ exports.updateDietByDietId = asyncHandler(async (req, res, next) => {
 // @access    Private
 exports.deleteDietByDietId = asyncHandler(async (req, res, next) => {
   const id = req.params.diet_log_id;
+
+  // Validate User Owns Resource
+  // const checkFitnessLog = await prisma.diet_logs.findUnique({
+  //   where: {
+  //     id: id,
+  //   },
+  // });
+
+  // if (checkFitnessLog.user_id !== req.user.id) {
+  //   return next(
+  //     new ErrorResponse('Not authorized to update this resource', 403)
+  //   );
+  // }
 
   const deletedDietLog = await prisma.diet_logs.delete({
     where: {
@@ -221,6 +290,19 @@ exports.createMeal = asyncHandler(async (req, res, next) => {
 exports.updateMealByMealId = asyncHandler(async (req, res, next) => {
   const id = req.params.meal_id;
 
+  // Validate User Owns Resource
+  const checkMeal = await prisma.meals.findUnique({
+    where: {
+      id: id,
+    },
+  });
+
+  if (checkMeal.user_id !== req.user.id) {
+    return next(
+      new ErrorResponse('Not authorized to update this resource', 403)
+    );
+  }
+
   const updatedMeal = await prisma.meals.update({
     where: {
       id: id,
@@ -246,6 +328,19 @@ exports.updateMealByMealId = asyncHandler(async (req, res, next) => {
 // @access    Private
 exports.deleteMealByMealId = asyncHandler(async (req, res, next) => {
   const id = req.params.meal_id;
+
+  // Validate User Owns Resource
+  const checkMeal = await prisma.meals.findUnique({
+    where: {
+      id: id,
+    },
+  });
+
+  if (checkMeal.user_id !== req.user.id) {
+    return next(
+      new ErrorResponse('Not authorized to update this resource', 403)
+    );
+  }
 
   const deletedMeal = await prisma.meals.delete({
     where: {
